requirements.txt
fastapi>=0.95.0,<0.96.0
uvicorn>=0.22.0,<0.23.0
pydantic>=1.10.7,<2.0.0
python-jose>=3.3.0,<4.0.0
passlib>=1.7.4,<2.0.0
python-multipart>=0.0.6,<0.1.0
loguru>=0.7.0,<0.8.0
slowapi>=0.1.7,<0.2.0
prometheus-fastapi-instrumentator>=6.0.0,<7.0.0
httpx>=0.24.0,<0.25.0
ruff>=0.0.262,<0.1.0
pytest>=7.3.1,<8.0.0
pytest-cov>=4.1.0,<5.0.0
boto3>=1.28.0,<2.0.0
